requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-organizations>=2.0.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
kombu>=5.3.0
gunicorn>=21.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.0
django-cors-headers>=4.3.0